- added new functionality to find bad etf to fix - changed yahoo downloader to add crunch numbers to requests - started playing with chaos - added scripts in python to use chaotic growth functions in trades with chposen etf - added scripts for basic NNs - added basic choosing investors - added new c code to simulate irational emotionaly based decisions = changed to volume trading - added loading btc data - added fractals generator - etc

+++ ETF_EVOL_DATA_ANALYZER/src/org/etf/analyzer/AnalyzePreference.java
@@ -216,7 +216,7 @@ public class AnalyzePreference
     private static void analyzeListOfLogFiles(File analyzedData,
 	    List<Path> expData)
     {
-	Map<String, Integer> holdMap = new HashMap<>();
+	Map<String, Float> holdMap = new HashMap<>();
 	Map<Integer, Map<String, Float>> holder = new HashMap<>();
 	Map<Integer, Map<String, AtomicInteger>> counterHolder = new HashMap<>();
 
@@ -251,7 +251,7 @@ public class AnalyzePreference
 	}
 
 	keySet.forEach(key -> {
-	    int holding = holdMap.get(key) == null ? 0 : holdMap.get(key);
+	    float holding = holdMap.get(key) == null ? 0 : holdMap.get(key);
 	    String row = key + ";"
 		    + holder.get(0).get(key) + ";"
 		    + holding + ";" + "http://www.etf.com/" + key
@@ -270,7 +270,7 @@ public class AnalyzePreference
 	});
     }
 
-    private static void extractPreferenceMaps(Map<String, Integer> holdMap,
+    private static void extractPreferenceMaps(Map<String, Float> holdMap,
 	    Map<Integer, Map<String, Float>> holder,
 	    Map<Integer, Map<String, AtomicInteger>> counterHolder, Path p)
 	    throws UnsupportedEncodingException, IOException
@@ -331,7 +331,7 @@ public class AnalyzePreference
 	    Arrays.asList(splitHold).forEach(held -> {
 		String[] parts = held.split("=");
 		String value = parts[1];
-		Integer holdVal = Integer.valueOf(value);
+		Float holdVal = Float.valueOf(value);
 		String key = parts[0];
 		if (holdMap.containsKey(key))
 		{
